backend/requirements.txt
# FastAPI ve temel bağımlılıklar
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Veritabanı
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# AI ve NLP
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
spacy==3.7.2

# Dosya işleme
Pillow==10.1.0
python-magic==0.4.27

# Test
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
coverage==7.3.2

# Güvenlik ve performans
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS ve middleware
fastapi-cors==0.0.6 